Bug fix: the error template is not being rendered with the proper token values, so when a runtime exception occurrs the rendered page does not have the proper header color, name of logged in user, etc. This is actually a problem with Dancer.
The solution was to patch Dancer/Error.pm so that it calls the
'before_error_render' hook with the same set of arguments that it
passes to 'before_template_render'. The second argument is a reference
to the hash of token values, which allows the hook to add, delete, and
modify tokens. I forked the Dancer repository, made this change, and
created a pull request. But who knows whether the maintainer will accept
it, or how long that will take. So I created a patch file, and added
two lines to Dockerfile. The first is a call to perl to determine the
full pathname of Dancer/Error.pm, and the second applies this patch.

This is an ugly hack, but there is no other feasible way of fixing this
bug. Fortunately, it is not likely that Dancer/Error.pm will ever be
modified to such a great extent that the patch will subsequently fail.
With any luck, the patch will be added to a subsequent release of
Dancer.

Web.pm now installs a 'before_error_render' hook that adds the tokens
'pbdb_site' and 'current_user' to the hash that will be used to render
the error template. This also requires a change to Classic to stash
the value of $user in a way that is retrievable by the hook.