Enhancement: additional debugging options
In order to debug slow actions, twree new settings are now recognized in pbdb.conf:
LOG_REQUESTS, PROFILE_REQUESTS, and PROFILE_THRESHOLD. Each can be set for a single
execution by setting an environment variable of the same name.

If LOG_REQUESTS is given a true value, lines will be appended to the file 'request_log'
in the classic log directory. This file will be created if it does not already exist.
Each request will generate a START line indicating the time, the action, and the pid
of the worker process that is handling it. A successful request will generate a
DONE line that indicates how many seconds the request took. A request that
crashes will generate an ERROR line.

If PROFILE_REQUESTS is given a true value and the method 'enable_profile' is available
in the DB package, and if the environment variable 'NYTPROF' is set with a value that
includes 'start=no', then &classic_request will bracket the request processing with calls
to DB::enable_profile and DB::finish_profile (via profile_request and profile_end_request
in Debug.pm). This enables the use of Devel::NYTProf to profile any given request. The
profile output will be written to a file called nytprof.<pid>.<action>.out.

The PROFILE_THRESHOLD setting defaults to 10 seconds. Any request whose execution time
exceeds that value will create a profile output file as described above. For each one,
a corresponding file whose name starts with 'request' will store the request parameters,
session, method, etc.

The command-line interface to debug_web.psgi now has additional options. You can run it
with the argument 'list', which will list all saved requests that match the specified
argument, or will list all saved requests with no argument. The argument can either be
an action name or a save file name. The suffix '.txt' is optional. The argument 'show'
will display the saved parameters and other information about the specified request.