Bug fix: When editing occurrence lists, sometimes an error would be generated for rows lacking a collection_no. Fixing this involved removing the data entry setting 'editable_collection_no', which appears to no longer be useful, and deleting the <div hide="editable_collection_no"></div> lines from occurrence_edit_row.html and reid_edit_row.html.
In Classic.pm at line 4926, I also added code to delete the fields 'occurrence_no'
and 'reid_no' with values of -1. The call to insertRecord immediately following this
line was throwing a 'column out of range' exception on dev.paleobiodb.org, probably
because the database on this server is set to strict mode. Deleting these fields
prevents this exception, and doesn't change the behavior.